Allow launching .desktop files or commands
Wofi is supposed to allow doing both from the same instance by
specifying `--show drun,run`, but it doesn't seem to work, so I'm
binding a separate key combo for each instead.

I also changed the keys for the launcher to `$mod+space`, which
conflicts with the default floating key combos, so I changed those.