Add key bindings to take screenshots
Following the instructions from this tutorial:

https://llandy3d.github.io/sway-on-ubuntu/extra/#screenshots

Note that the explicit path to `grimshot` has to be given since
`~/.local/bin/` is added to the `PATH` in my `.bashrc`, which hasn't
been sourced before sway is started by the display manager. Hopefully
running sway without a display manager will eventually fix that.